Co-locate summary diagram information
Co-locate summary diagram information. Make the public structs private
since they are only needed for diagram assembly.
